# src/main/resources/yapi_export_api/swaggerApi.json
{
  "swagger": "2.0",
  "info": {
    "title": "ERP-METRICS",
    "version": "last",
    "description": "erp-metrics 雷神架构组项目"
  },
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/app/monitor/report/reportByApp": {
      "post": {
        "tags": [
          "应用异常监控报表相关的接口"
        ],
        "summary": "按应用维度统计",
        "description": "按应用维度统计",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "root",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "deptId": {
                  "type": "integer",
                  "description": "归属项目部的ID"
                },
                "appOwner": {
                  "type": "string",
                  "description": "应用负责人"
                },
                "appId": {
                  "type": "integer",
                  "description": "应用的ID"
                },
                "startDate": {
                  "type": "string",
                  "description": "开始日期"
                },
                "endDate": {
                  "type": "string",
                  "description": "结束日期"
                },
                "status": {
                  "type": "integer",
                  "description": "状态：0表示未完成，1表示完成"
                }
              },
              "$schema": "http://json-schema.org/draft-04/schema#"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string",
              "format": "binary",
              "default": "{\n\"msg\": \"成功\",\n\"result\": 1,\n\"data\": {\n\"雷神项目部\": [\n{\n\"deptId\": \"1\",\n\"deptName\": \"雷神项目部\",\n\"appOwner\": \"江超\",\n\"appId\": \"18\",\n\"appName\": \"ld-erp-hr线上应用异常监控\",\n\"createDate\": \"2021-01-10\",\n\"times\": 4,\n\"total\": 21\n},\n{\n\"deptId\": \"1\",\n\"deptName\": \"雷神项目部\",\n\"appOwner\": \"江超\",\n\"appId\": \"18\",\n\"appName\": \"ld-erp-hr线上应用异常监控\",\n\"createDate\": \"2021-01-09\",\n\"times\": 3,\n\"total\": 18\n},\n{\n\"deptId\": \"1\",\n\"deptName\": \"雷神项目部\",\n\"appOwner\": \"江超\",\n\"appId\": \"14\",\n\"appName\": \"erp-gateway线上应用异常监控\",\n\"createDate\": \"2021-01-08\",\n\"times\": 2,\n\"total\": 30\n},\n{\n\"deptId\": \"1\",\n\"deptName\": \"雷神项目部\",\n\"appOwner\": \"江超\",\n\"appId\": \"16\",\n\"appName\": \"erp-message-push线上应用异常监控\",\n\"createDate\": \"2021-01-08\",\n\"times\": 1,\n\"total\": 3\n},\n{\n\"deptId\": \"1\",\n\"deptName\": \"雷神项目部\",\n\"appOwner\": \"江超\",\n\"appId\": \"18\",\n\"appName\": \"ld-erp-hr线上应用异常监控\",\n\"createDate\": \"2021-01-08\",\n\"times\": 4,\n\"total\": 28\n},\n{\n\"deptId\": \"1\",\n\"deptName\": \"雷神项目部\",\n\"appOwner\": \"江超\",\n\"appId\": \"44\",\n\"appName\": \"erp-user线上应用异常监控\",\n\"createDate\": \"2021-01-08\",\n\"times\": 1,\n\"total\": 10\n},\n{\n\"deptId\": \"1\",\n\"deptName\": \"雷神项目部\",\n\"appOwner\": \"江超\",\n\"appId\": \"18\",\n\"appName\": \"ld-erp-hr线上应用异常监控\",\n\"createDate\": \"2021-01-07\",\n\"times\": 4,\n\"total\": 21\n},\n{\n\"deptId\": \"1\",\n\"deptName\": \"雷神项目部\",\n\"appOwner\": \"江超\",\n\"appId\": \"18\",\n\"appName\": \"ld-erp-hr线上应用异常监控\",\n\"createDate\": \"2021-01-06\",\n\"times\": 1,\n\"total\": 7\n}\n]\n}\n}"
            }
          }
        }
      }
    },
    "/app/monitor/report/reportByOwner": {
      "post": {
        "tags": [
          "应用异常监控报表相关的接口"
        ],
        "summary": "按应用负责人维度统计",
        "description": "按应用负责人维度统计",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "root",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "deptId": {
                  "type": "integer",
                  "description": "归属项目部的ID"
                },
                "appOwner": {
                  "type": "string",
                  "description": "应用负责人"
                },
                "startDate": {
                  "type": "string",
                  "description": "开始日期，格式yyyy-MM-dd"
                },
                "endDate": {
                  "type": "string",
                  "description": "结束日期，格式yyyy-MM-dd"
                },
                "status": {
                  "type": "integer",
                  "description": "状态：0表示未完成，1表示完成"
                }
              },
              "$schema": "http://json-schema.org/draft-04/schema#"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string",
              "format": "binary",
              "default": "{\n\"msg\": \"成功\",\n\"result\": 1,\n\"data\": {\n\"雷神项目部\": [\n{\n\"deptId\": \"1\",\n\"deptName\": \"雷神项目部\",\n\"appOwner\": \"江超\",\n\"createDate\": \"2021-01-10\",\n\"times\": 4,\n\"total\": 21\n},\n{\n\"deptId\": \"1\",\n\"deptName\": \"雷神项目部\",\n\"appOwner\": \"江超\",\n\"createDate\": \"2021-01-09\",\n\"times\": 3,\n\"total\": 18\n},\n{\n\"deptId\": \"1\",\n\"deptName\": \"雷神项目部\",\n\"appOwner\": \"江超\",\n\"createDate\": \"2021-01-08\",\n\"times\": 8,\n\"total\": 71\n},\n{\n\"deptId\": \"1\",\n\"deptName\": \"雷神项目部\",\n\"appOwner\": \"江超\",\n\"createDate\": \"2021-01-07\",\n\"times\": 4,\n\"total\": 21\n},\n{\n\"deptId\": \"1\",\n\"deptName\": \"雷神项目部\",\n\"appOwner\": \"江超\",\n\"createDate\": \"2021-01-06\",\n\"times\": 1,\n\"total\": 7\n}\n]\n}\n}"
            }
          }
        }
      }
    },
    "/app/monitor/report/reportByDept": {
      "post": {
        "tags": [
          "应用异常监控报表相关的接口"
        ],
        "summary": "按归属项目部维度统计",
        "description": "按归属项目部维度统计",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "root",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "deptId": {
                  "type": "integer"
                },
                "startDate": {
                  "type": "string",
                  "description": "开始日期，格式yyyy-MM-dd"
                },
                "endDate": {
                  "type": "string",
                  "description": "结束日期，格式yyyy-MM-dd"
                },
                "status": {
                  "type": "integer",
                  "description": "状态：0表示未完成，1表示完成"
                }
              },
              "$schema": "http://json-schema.org/draft-04/schema#"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string",
              "format": "binary",
              "default": "{\n\"msg\": \"成功\",\n\"result\": 1,\n\"data\": {\n\"雷神项目部\": [\n{\n\"deptId\": \"1\",\n\"deptName\": \"雷神项目部\",\n\"createDate\": \"2021-01-10\",\n\"times\": 12,\n\"total\": 32\n},\n{\n\"deptId\": \"1\",\n\"deptName\": \"雷神项目部\",\n\"createDate\": \"2021-01-09\",\n\"times\": 12,\n\"total\": 32\n},\n{\n\"deptId\": \"1\",\n\"deptName\": \"雷神项目部\",\n\"createDate\": \"2021-01-08\",\n\"times\": 84,\n\"total\": 395\n},\n{\n\"deptId\": \"1\",\n\"deptName\": \"雷神项目部\",\n\"createDate\": \"2021-01-07\",\n\"times\": 62,\n\"total\": 207\n},\n{\n\"deptId\": \"1\",\n\"deptName\": \"雷神项目部\",\n\"createDate\": \"2021-01-06\",\n\"times\": 32,\n\"total\": 93\n}\n]\n}\n}"
            }
          }
        }
      }
    },
    "/app/monitor/report/reportByDate": {
      "post": {
        "tags": [
          "应用异常监控报表相关的接口"
        ],
        "summary": "按日期维度统计",
        "description": "按日期维度统计",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "root",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "startDate": {
                  "type": "string",
                  "description": "开始时间，格式yyyy-MM-dd"
                },
                "endDate": {
                  "type": "string",
                  "description": "结束时间，格式yyyy-MM-dd"
                },
                "status": {
                  "type": "integer",
                  "description": "状态：0表示未完成，1表示完成"
                }
              },
              "$schema": "http://json-schema.org/draft-04/schema#"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string",
              "format": "binary",
              "default": "{\n\"msg\": \"成功\",\n\"result\": 1,\n\"data\": [\n{\n\"createDate\": \"2021-01-10\",\n\"times\": 12,\n\"total\": 32\n},\n{\n\"createDate\": \"2021-01-09\",\n\"times\": 12,\n\"total\": 32\n},\n{\n\"createDate\": \"2021-01-08\",\n\"times\": 84,\n\"total\": 395\n},\n{\n\"createDate\": \"2021-01-07\",\n\"times\": 62,\n\"total\": 207\n},\n{\n\"createDate\": \"2021-01-06\",\n\"times\": 32,\n\"total\": 93\n}\n]\n}"
            }
          }
        }
      }
    },
...
  }
}